% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_map.R
\name{ab_map}
\alias{ab_map}
\title{The function maps data from one set to another using a lookup table}
\usage{
ab_map(dat, map)
}
\arguments{
\item{dat}{vector to be mapped}

\item{map}{data.frame with two columns defining the mapping}
}
\description{
The function maps data from one set to another using a lookup table
}
\examples{
ab_map(
  c(1,2,3,1,2,3),
  data.frame(c(1, 2, 3), c("cat", "dog", "dog"))
)
}
