% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_label_encoding.R
\name{apply_label_encoding}
\alias{apply_label_encoding}
\title{Apply Label Encoding to Data}
\usage{
apply_label_encoding(data)
}
\arguments{
\item{data}{A data frame where some columns may be non-numeric.}
}
\value{
A data frame with all columns converted to numeric type,
where non-numeric columns have been label encoded.
}
\description{
This function iterates through each column in a data frame.
If a column is not numeric, it applies label encoding,
converting categories into integers. A progress bar is displayed
showing the progress and the column currently being processed.
}
