% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Clean Data by Removing Columns with Excessive Missing Values}
\usage{
clean_data(data, m)
}
\arguments{
\item{data}{A data frame that may contain missing values.}

\item{m}{The maximum allowable proportion of missing values for a column to be retained.}
}
\value{
A data frame with columns having a higher proportion of missing values than the specified threshold removed.
}
\description{
This function examines each column in a data frame and removes columns where the proportion of missing values exceeds a specified threshold.
}
