% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_files.R
\name{combine_files}
\alias{combine_files}
\title{Combine multiple excel files with identical fields into a master data set}
\usage{
combine_files()
}
\arguments{
\item{dir}{the directory where the excel files live}
}
\description{
Combine multiple excel files with identical fields into a master data set
}
