% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_aidv.R
\name{generate_aidv}
\alias{generate_aidv}
\title{Generate an aidv file for EVS}
\usage{
generate_aidv(data, base_elevation, elevation_units, result_units, fp_out)
}
\arguments{
\item{data}{data.frame to be converted to aidv format for use in EVS; names(data)==c("x", "y", "ztop",  "xbot", "result_name_1", "result_name_2", ... ,"boring")}

\item{base_elevation}{ground elevation at sample location}

\item{elevation_units}{elevation units; defaults to "ft"}

\item{result_units}{vector length must be equal to the number of analytes}

\item{fp_out}{file path to output aidv file}
}
\description{
Generate an aidv file for EVS
}
\examples{
data <- data.frame(
   x = c(1,2),
   y = c(3,4),
   ztop = c(5, 6),
   zbot = c(3, 2),
   Analyte1 = c(2, 1),
   Analyte2 = c(3, 5),
   boring = c("a1", "b1"),
   surface_elevation = c(0, 1)
)
base_elevation <- 4
elevation_units <- "feet"
result_units <- c("ug/l", "ug/l")
fp_out <- "my_data.aidv"
generate_aidv(data, base_elevation, elevation_units="ft", result_units, fp_out)
}
