% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_apdv.R
\name{generate_apdv}
\alias{generate_apdv}
\title{Generate apdv file for EVS}
\usage{
generate_apdv(
  data,
  base_elevation,
  elevation_units = "ft",
  result_units,
  fp_out
)
}
\arguments{
\item{data}{data.frame to be converted to aidv format for use in EVS; names(data)==c("x", "y", "z", "result_name_1", "result_name_2", ... ,"boring")}

\item{base_elevation}{ground elevation at sample location}

\item{elevation_units}{elevation units; defaults to "ft"}

\item{result_units}{vector length must be equal to the number of analytes}

\item{fp_out}{file path to output apdv file}
}
\description{
Generate apdv file for EVS
}
\examples{
data <- data.frame(
   x = c(1,2),
   y = c(3,4),
   z = c(5, 6),
   Analyte1 = c(2, 1),
   Analyte2 = c(3, 5),
   boring = c("a1", "b1"),
   surface_elevation = c(0, 1)
)
base_elevation <- 4
elevation_units <- "feet"
result_units <- c("ug/l", "ug/l")
fp_out <- "my_data.apdv"
generate_apdv(data, base_elevation, elevation_units, result_units, fp_out)
}
