% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dem.R
\name{get_dem}
\alias{get_dem}
\title{Pull in DEM from Amazon Web Services}
\usage{
get_dem(df, epsg)
}
\arguments{
\item{df}{data.frame of x and y coordinates}

\item{epsg}{epsg code for coordinate pair}
}
\description{
Pull in DEM from Amazon Web Services
}
