% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dem.R
\name{get_elevation_point}
\alias{get_elevation_point}
\title{Get elevation at a set of points using Amazon Web Services}
\usage{
get_elevation_point(df, epsg)
}
\arguments{
\item{df}{data.frame of x and y coordinates}

\item{epsg}{epsg code for coordinate pair}
}
\description{
Get elevation at a set of points using Amazon Web Services
}
