% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_utilities.R
\name{get_month}
\alias{get_month}
\title{Function returns month associated with a POSIXct date}
\usage{
get_month(date)
}
\arguments{
\item{date}{POSIXct date}
}
\description{
Function returns month associated with a POSIXct date
}
