% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_utilities.R
\name{get_year}
\alias{get_year}
\title{Function returns year associated with a POSIXct data}
\usage{
get_year(date)
}
\arguments{
\item{date}{POSIXct date}
}
\description{
Function returns year associated with a POSIXct data
}
