% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_binary_columns.R
\name{identify_binary_columns}
\alias{identify_binary_columns}
\title{Identify Binary Columns}
\usage{
identify_binary_columns(data)
}
\arguments{
\item{data}{A data frame that may contain binary columns.}
}
\value{
A vector containing the names of all columns that are binary (only containing values 0 and 1).
}
\description{
This function iterates through each column in a data frame. It identifies columns that contain only the values 0 and 1,
which are considered binary columns. It then returns the names of these binary columns.
It includes a progress bar to indicate the progress of identification.
}
