% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jcrypt.R
\name{jcrypt}
\alias{jcrypt}
\title{Encrypt or decrypt a character vector}
\usage{
jcrypt(vec, key, encrypt = T)
}
\arguments{
\item{vec}{character vector to encrypt}

\item{key}{encryption/decrytion key}

\item{encrypt}{encrypts when encrypt==T; decrypts when encrypt==F}
}
\description{
Encrypt or decrypt a character vector
}
