% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_cass.R
\name{lookup_cass}
\alias{lookup_cass}
\title{Function tried to find a cas number associated with a string using the common chemistry database}
\usage{
lookup_cass(search_string)
}
\arguments{
\item{search_string}{name of analyte}
}
\value{
if the string is associated with a cass number in the database, the cass number will be returned; if the string is not associated with a cass number, NA will be returned
}
\description{
Function tried to find a cas number associated with a string using the common chemistry database
}
