% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize a vector}
\usage{
normalize(x)
}
\arguments{
\item{x}{A numeric vector that you want to normalize.}
}
\value{
A numeric vector of the same length, with values normalized to the range 0 to 1.
}
\description{
This function takes a numeric vector and normalizes it to a range of 0 to 1.
It adds a small epsilon to the range to avoid division by zero.
It includes a progress bar to indicate the progress of normalization.
}
