% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_df.R
\name{normalize_df}
\alias{normalize_df}
\title{Normalize columns in a data frame}
\usage{
normalize_df(df)
}
\arguments{
\item{df}{A data frame with numeric and/or other type columns.}
}
\value{
A data frame with numeric columns normalized to the range 0 to 1, other columns remain unchanged.
}
\description{
This function takes a data frame and normalizes its numeric columns to a range of 0 to 1, ignoring NA values.
It calculates the range for each numeric column and scales the values accordingly.
For columns with a range of 0 (all values are the same), it returns a vector of zeros for non-NA values.
}
