% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{package_versions}
\alias{package_versions}
\title{Determine package versions.}
\usage{
package_versions(packages = NULL, base = FALSE, sort = FALSE)
}
\arguments{
\item{packages}{Packages to query; if not specified (or \code{NULL}),
all currently attached packages are queried.}

\item{base}{If \code{TRUE}, include the base package (R version).}

\item{sort}{If \code{TRUE}, sort package names alphabetically.}
}
\value{
Data frame with package names and corresponding versions.
}
\description{
Determine package versions.
}
