% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_to_xyz.R
\name{raster_to_xyz}
\alias{raster_to_xyz}
\title{Convert raster data to an xyz point cloud}
\usage{
raster_to_xyz(fp_raster, fp_out)
}
\arguments{
\item{fp_raster}{file path of input raster}

\item{fp_out}{file path of output pointcloud}
}
\description{
Convert raster data to an xyz point cloud
}
