% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_subset.R
\name{read_subset}
\alias{read_subset}
\title{Read a Subset of a Large Dataset}
\usage{
read_subset(fn, fn_fields, k = NA, p = NA, total = NA)
}
\arguments{
\item{fn}{The file name of the large dataset.}

\item{fn_fields}{The file name containing the field names.}

\item{k}{The number of rows to sample from the dataset (optional).}

\item{p}{The number of fields to sample from the dataset (optional).}

\item{total}{The total number of rows in the dataset (optional, will be calculated if not provided).}
}
\value{
A data frame containing the sampled subset of the dataset.
}
\description{
This function reads a subset of rows and/or columns from a large dataset.
It allows sampling of a specific number of rows (k) and/or a specific number of fields (p).
It's particularly useful when dealing with very large files where reading the entire file into memory is not feasible.
}
