% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_to_text.R
\name{result_to_text}
\alias{result_to_text}
\title{Function formats analytical chemistry results}
\usage{
result_to_text(RESULT, DETECT_FLAG, DETECT_LIMIT)
}
\arguments{
\item{RESULT}{numeric vector of analytical chemistry results}

\item{DETECT_FLAG}{vector that whether or not the result lies beneath a detection limit; possible values are "N" and "Y"}

\item{DETECT_LIMIT}{the detection limits associate with the samples}
}
\value{
returns a character vector compatible with EVS
}
\description{
Function formats analytical chemistry results
}
