% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_fields.R
\name{search_fields}
\alias{search_fields}
\title{Search and Select Specific Fields from Data}
\usage{
search_fields(data, name_list)
}
\arguments{
\item{data}{A data frame from which to select columns.}

\item{name_list}{A vector of base names to construct the search patterns from.}
}
\value{
A data frame containing only the columns that match the constructed search patterns.
}
\description{
This function searches through the column names of a data frame and selects columns that match a specified pattern.
The patterns are constructed from a list of base names provided by the user, aimed to match a specific naming convention.
}
