% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see_color.R
\name{see_col}
\alias{see_col}
\title{View a color or a set of color overlays}
\usage{
see_col(cols)
}
\arguments{
\item{cols}{a list of color overlays (see example)}
}
\description{
View a color or a set of color overlays
}
\examples{
see_col(
 list(
   rgb(0, 0, 0, 1),
   rgb(1, 0, 0, 0.2)
 )
)
}
