\name{summary.utility.endnode.parfun1d}

\alias{summary.utility.endnode.parfun1d}

\title{Print Summary of Node Definition}

\description{
Print summary of node definition.
}

\usage{
\method{summary}{utility.endnode.parfun1d}(object, ...)
}

\arguments{
  \item{object}{
node of which a summary is to be printed.
}
  \item{...}{
currently no other arguments are implemented or passed further.
}
}

\note{
In the current version of the package, the methods \code{print} and \code{summary}
provide the same output.
}

\references{
Short description of the package: \cr\cr
Reichert, P., Schuwirth, N. and Langhans, S., 
Constructing, evaluating and visualizing value and utility functions for decision support, submitted. \cr\cr
Textbooks on the use of utility and value functions in decision analysis: \cr\cr
Keeney, R. L. and Raiffa, H. Decisions with Multiple Objectives. Cambridge University Press, 1993. First published in 1976. \cr\cr
Eisenfuehr, F., Weber, M. and Langer, T., Rational Decision Making, Springer, Berlin, 2010.
}

\author{
Peter Reichert <peter.reichert@eawag.ch>
}

\seealso{
See \code{\link{utility.endnode.parfun1d.create}} for how to construct such a node.
}

\examples{
# see
help(utility)
# for examples.
}
