\name{utility.get.attrib.names}

\alias{utility.get.attrib.names}

\title{Get Names of Attributes Used by a Value Functions}

\description{Function to get the names of the attributes used by a given value function.}

\usage{
utility.get.attrib.names(node)
}

\arguments{
  \item{node}{
node of an objectives hierarchy with its associated value function (object of class \code{utility}).
}
}

\value{
Character vector of names of attributes.
}

\references{
Short description of the package: \cr\cr
Reichert, P., Schuwirth, N. and Langhans, S., 
Constructing, evaluating and visualizing value and utility functions for decision support, Environmental Modelling & Software 46, 283-291, 2013. \cr\cr
Textbooks on the use of utility and value functions in decision analysis: \cr\cr
Keeney, R. L. and Raiffa, H. Decisions with Multiple Objectives - Preferences and Value Tradeoffs. John Wiley & Sons, 1976. \cr\cr
Eisenfuehr, F., Weber, M. and Langer, T., Rational Decision Making, Springer, Berlin, 2010.
}

\author{
Peter Reichert <peter.reichert@eawag.ch>
}


