\name{utility.structure}

\alias{utility.structure}

\title{Extract Structure of Objectives Hierarchy}

\description{Function to extract the structure of an objectives hierarchy.}

\usage{
utility.structure(node)
}

\arguments{
  \item{node}{
object containing the utility or value function.
}
}

\value{
Data frame containing structural information of the objectives hierarchy.
}

\references{
Short description of the package: \cr\cr
Reichert, P., Schuwirth, N. and Langhans, S., 
Constructing, evaluating and visualizing value and utility functions for decision support, Environmental Modelling & Software 46, 283-291, 2013. \cr\cr
Textbooks on the use of utility and value functions in decision analysis: \cr\cr
Keeney, R. L. and Raiffa, H. Decisions with Multiple Objectives - Preferences and Value Tradeoffs. John Wiley & Sons, 1976. \cr\cr
Eisenfuehr, F., Weber, M. and Langer, T., Rational Decision Making, Springer, Berlin, 2010.
}

\author{
Peter Reichert <peter.reichert@emeriti.eawag.ch>
}


