% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_from_raster.R
\name{z_from_topo}
\alias{z_from_topo}
\title{Extract value from raster at a specific spatial location}
\usage{
z_from_topo(x, y, topo)
}
\arguments{
\item{x}{x coordinate}

\item{y}{y coordinate}

\item{z}{raster object}
}
\description{
Extract value from raster at a specific spatial location
}
